pyproject.toml
[project]
name = "weather-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "amit",email = "pandheramit245@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.61.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "flask-cors (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
